Move remaining blockdev func unit tests to disk unit tests and delete file
