+++ tests/unit/modules/disk_test.py
@@ -7,12 +7,15 @@
 from __future__ import absolute_import
 
 # Import Salt Testing libs
-from salttesting import TestCase
+from salttesting import TestCase, skipIf
 from salttesting.helpers import ensure_in_syspath
 from salttesting.mock import MagicMock, patch
 ensure_in_syspath('../../')
 
+# Import Salt libs
 from salt.modules import disk
+import salt.utils
+
 #usage_size = {'filesystem': None,'1K-blocks':10000,'used':10000,'available':10000,'capacity':10000}
 
 STUB_DISK_USAGE = {
@@ -129,6 +132,27 @@ class DiskTestCase(TestCase):
                     'blockdev --setra 512 --setfra 512 /dev/sda',
                     python_shell=False
                 )
+
+    @skipIf(not salt.utils.which('sync'), 'sync not found')
+    @skipIf(not salt.utils.which('mkfs'), 'mkfs not found')
+    def test_format(self):
+        '''
+        unit tests for disk.format
+        '''
+        device = '/dev/sdX1'
+        mock = MagicMock(return_value=0)
+        with patch.dict(disk.__salt__, {'cmd.retcode': mock}):
+            self.assertEqual(disk.format_(device), True)
+
+    def test_fstype(self):
+        '''
+        unit tests for disk.fstype
+        '''
+        device = '/dev/sdX1'
+        fs_type = 'ext4'
+        mock = MagicMock(return_value='FSTYPE\n{0}'.format(fs_type))
+        with patch.dict(disk.__salt__, {'cmd.run': mock}):
+            self.assertEqual(disk.fstype(device), fs_type)
 
 
 if __name__ == '__main__':
